refactor: 훅 종류별로 분류해 파일로 정리
