+++ src/hooks/useCalculateCouponDiscount.ts
@@ -1,6 +1,9 @@
-import { useRecoilValue } from 'recoil';
+import { useRecoilState, useRecoilValue } from 'recoil';
 
 import { selectedCartItemListState } from '../recoil/CartItem/atoms/selectedCartItemListState';
+import { selectedCartItemListTotalPriceSelector } from '../recoil/CartItem/selectors/selectedCartItemListTotalPriceSelector';
+import { selectedCouponListState } from '../recoil/Coupon/atoms/selectedCouponListState';
+import { selectedCouponTotalDiscountState } from '../recoil/Coupon/atoms/selectedCouponTotalDiscountState';
 import { deliveryFeeState } from '../recoil/DeliveryFee/atoms/deliveryFeeState';
 import { Coupon } from '../types/Coupon.type';
 
@@ -34,5 +37,40 @@ export function useCalculateCouponDiscount() {
       }
     }
   };
+
   return { calculateCouponDiscount };
+}
+
+export function useCalculateTotalCouponDiscount() {
+  const selectedCartItemTotalPrice = useRecoilValue(selectedCartItemListTotalPriceSelector);
+  const { calculateCouponDiscount } = useCalculateCouponDiscount();
+  const [selectedCouponTotalDiscount, setSelectedCouponTotalDiscount] = useRecoilState(
+    selectedCouponTotalDiscountState,
+  );
+  const selectedCouponList = useRecoilValue(selectedCouponListState);
+
+  const calculateTotalCouponDiscount = () => {
+    const bogoCoupons = selectedCouponList.filter((coupon) => coupon.discountType === 'buyXgetY');
+    const bogoDiscountedPrice = bogoCoupons.reduce((acc, cur) => {
+      return acc - calculateCouponDiscount(acc, cur);
+    }, selectedCartItemTotalPrice);
+
+    const percentageCoupons = selectedCouponList.filter((coupon) => coupon.discountType === 'percentage');
+    percentageCoupons.sort((a, b) => b.discount! - a.discount!);
+    const percentageDiscountedPrice = percentageCoupons.reduce((acc, cur) => {
+      return acc - calculateCouponDiscount(acc, cur);
+    }, bogoDiscountedPrice);
+
+    const fixedCoupons = selectedCouponList.filter(
+      (coupon) => coupon.discountType !== 'percentage' && coupon.discountType !== 'buyXgetY',
+    );
+    const totalDiscountedPrice = fixedCoupons.reduce((acc, cur) => {
+      return acc - calculateCouponDiscount(acc, cur);
+    }, percentageDiscountedPrice);
+
+    const totalCouponDiscount = selectedCartItemTotalPrice - totalDiscountedPrice;
+    setSelectedCouponTotalDiscount(totalCouponDiscount);
+  };
+
+  return { selectedCouponTotalDiscount, calculateTotalCouponDiscount };
 }
