+++ src/hooks/useCouponCheck.ts
@@ -3,15 +3,23 @@ import { useRecoilValue } from 'recoil';
 import { selectedCouponListState } from '../recoil/Coupon/atoms/selectedCouponListState';
 import { Coupon } from '../types/Coupon.type';
 import { useCalculateDeliveryFee } from './useCalculateDeliveryFee';
-import { useCouponValidator } from './useCouponValidator';
 
-export const useCouponApplicabilityChecker = () => {
-  const { isCouponValid } = useCouponValidator();
+export const useCouponCheck = () => {
   const { deliveryFee } = useCalculateDeliveryFee();
   const selectedCouponList = useRecoilValue(selectedCouponListState);
 
+  const isCouponExpired = (expirationDate: string) => {
+    const today = new Date();
+    const expiration = new Date(expirationDate);
+    return expiration < today;
+  };
+
+  const isCouponValid = (coupon: Coupon) => {
+    return !isCouponExpired(coupon.expirationDate);
+  };
+
   const isCouponApplicable = (coupon: Coupon, totalAmount: number, now: Date = new Date()) => {
-    if (selectedCouponList.length === 2 && !selectedCouponList.find((item) => item.id == coupon.id)) return false;
+    if (selectedCouponList.length === 2 && !selectedCouponList.find((item) => item.id === coupon.id)) return false;
     if (!coupon || !isCouponValid(coupon)) return false;
 
     if (coupon.minimumAmount && totalAmount < coupon.minimumAmount) {
@@ -20,11 +28,9 @@ export const useCouponApplicabilityChecker = () => {
 
     if (coupon.availableTime) {
       const [startHour, startMinute, startSecond] = coupon.availableTime.start.split(':').map(Number);
-
       const [endHour, endMinute, endSecond] = coupon.availableTime.end.split(':').map(Number);
 
       const startTime = new Date(now.getFullYear(), now.getMonth(), now.getDate(), startHour, startMinute, startSecond);
-
       const endTime = new Date(now.getFullYear(), now.getMonth(), now.getDate(), endHour, endMinute, endSecond);
 
       if (now < startTime || now > endTime) {
@@ -40,6 +46,7 @@ export const useCouponApplicabilityChecker = () => {
   };
 
   return {
+    isCouponValid,
     isCouponApplicable,
   };
 };
