--- src/hooks/useCouponValidator.ts
@@ -1,17 +1,0 @@
-import { Coupon } from '../types/Coupon.type';
-
-export const useCouponValidator = () => {
-  const isCouponExpired = (expirationDate: string) => {
-    const today = new Date();
-    const expiration = new Date(expirationDate);
-    return expiration < today;
-  };
-
-  const isCouponValid = (coupon: Coupon) => {
-    return !isCouponExpired(coupon.expirationDate);
-  };
-
-  return {
-    isCouponValid,
-  };
-};
