+++ src/pages/PaymentConfirmPage/PaymentConfirmPage.tsx
@@ -5,8 +5,8 @@ import SubmitButton from '../../components/Button/SubmitButton/SubmitButton';
 import TitleContainer from '../../components/Container/TitleContainer/TitleContainer';
 import Header from '../../components/Header/Header';
 import { PATHS } from '../../constants/PATHS';
+import { useCalculateTotalCouponDiscount } from '../../hooks/useCalculateCouponDiscount';
 import { useCalculateDeliveryFee } from '../../hooks/useCalculateDeliveryFee';
-import { useCalculateTotalCouponDiscount } from '../../hooks/useCalculateTotalCouponDiscount';
 import { selectedCartItemListState } from '../../recoil/CartItem/atoms/selectedCartItemListState';
 import { selectedCartItemListTotalCountSelector } from '../../recoil/CartItem/selectors/selectedCartItemListTotalCountSelector';
 import { selectedCartItemListTotalPriceSelector } from '../../recoil/CartItem/selectors/selectedCartItemListTotalPriceSelector';
